Fix page title property
When creating a Page via the API, the docs aren't very clear about how
the title should be set in the Page's properties. The structure is
different than other properties only in that there's no higher-level
property-name for the title, only the "title" key as an object of a
property with a text content.