Return images from paragraphs
Along with the rich_text that would be derived via conversion of the
span-level elements, we must return any image that was inlined. Most
images appear inside a paragraph anyway so this was unavoidable. We also
update image links to return their images as well. This imposes a
requirement to return any block-level elements resulting from conversion
as the first element in a tuple (2-element list).