Add Notion Decorator
Processes the parsed Border and generates a page and a function to
generate the subpages. This is because the subpages need to be created
with a page id. This gives the user the responsibility to call the
function after creating the parent page, but it's better than having the
user of the class map over the list of subpages to set the id on each
one after creating the parent page.